Revert heavy RTL selector ESLint configuration
