+++ x-pack/solutions/security/packages/upselling/.eslintrc.js
@@ -1,10 +1,8 @@
 /*
  * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the "Elastic License
- * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
- * Public License v 1"; you may not use this file except in compliance with, at
- * your election, the "Elastic License 2.0", the "GNU Affero General Public
- * License v3.0 only", or the "Server Side Public License, v 1".
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
  */
 
 /**
@@ -70,6 +68,7 @@
 
 // eslint-disable-next-line import/no-nodejs-modules
 const path = require('path');
+// eslint-disable-next-line import/no-nodejs-modules
 const { execSync } = require('child_process');
 
 const { minimatch } = require('minimatch');
