--- x-pack/solutions/security/plugins/security_solution/public/.eslintrc.js
@@ -1,9 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-// eslint-disable-next-line @kbn/imports/no_boundary_crossing
-module.exports = require('../../../../../../packages/kbn-eslint-config-security-tests')(__dirname);
